api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
appnope==0.1.0            # via ipython
babel==2.6.0              # via flower
backcall==0.1.0           # via ipython
billiard==3.5.0.4         # via celery
cachetools==3.0.0         # via google-auth
celery[gevent]==4.2.1
certifi==2018.10.15       # via requests
chardet==3.0.4            # via requests
coverage==4.5.2           # via coveralls
coveralls==1.5.1
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid, social-auth-core
django-cors-headers==2.4.0
django-environ==0.4.5
django-registration-redux==2.5
django-storages[google]==1.7.1
django==2.1.3
djangorestframework-csv==2.1.0
djangorestframework-link-header-pagination==0.1.1
djangorestframework==3.9.0
docopt==0.6.2             # via coveralls
flower==0.9.2
fuzzywuzzy[speedup]==0.17.0
gevent==1.3.7             # via celery
google-api-core==1.5.2    # via google-cloud-core, google-cloud-storage
google-api-python-client==1.7.4
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.1        # via google-api-core, google-api-python-client, google-auth-httplib2
google-cloud-core==0.28.1  # via google-cloud-storage
google-cloud-storage==1.13.0  # via django-storages
google-resumable-media==0.3.1  # via google-cloud-storage
googleapis-common-protos==1.5.5  # via google-api-core
greenlet==0.4.15          # via gevent
gunicorn==19.9.0
httplib2==0.12.0          # via google-api-python-client, google-auth-httplib2
idna==2.7                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.1.1
jedi==0.13.1              # via ipython
jellyfish==0.6.1          # via recordlinkage
kombu==4.2.1              # via celery
numpy==1.15.4             # via pandas, recordlinkage, scikit-learn, scipy
oauthlib==2.1.0           # via requests-oauthlib, social-auth-core
pandas==0.23.4
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==2.0.7     # via ipython
protobuf==3.6.1           # via google-api-core, googleapis-common-protos
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0         # via pexpect
pyasn1-modules==0.2.2     # via google-auth
pyasn1==0.4.4             # via pyasn1-modules, rsa
pygments==2.3.0           # via ipython
pyjwt==1.6.4              # via social-auth-core
python-dateutil==2.7.5    # via pandas
python-levenshtein==0.12.0  # via fuzzywuzzy
python-magic==0.4.15
python3-openid==3.1.0     # via social-auth-core
pytz==2018.7              # via babel, celery, django, flower, google-api-core, pandas
raven==6.9.0
recordlinkage==0.12
redis==2.10.6
requests-oauthlib==1.0.0  # via social-auth-core
requests==2.20.1          # via coveralls, google-api-core, requests-oauthlib, social-auth-core
rsa==4.0                  # via google-auth
scikit-learn==0.20.1      # via recordlinkage
scipy==1.1.0              # via recordlinkage, scikit-learn
six==1.11.0               # via djangorestframework-csv, google-api-core, google-api-python-client, google-auth, google-resumable-media, prompt-toolkit, protobuf, python-dateutil, recordlinkage, social-auth-app-django, social-auth-core, traitlets
social-auth-app-django==3.1.0
social-auth-core==2.0.0   # via social-auth-app-django
tornado==4.5.3
traitlets==4.3.2          # via ipython
unicodecsv==0.14.1        # via djangorestframework-csv
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.24.1           # via requests
vine==1.1.4               # via amqp
wcwidth==0.1.7            # via prompt-toolkit
